PHP 8.1 | Date_Helper: add input validation [1]
As of PHP 8.1, passing `null` to not explicitly nullable scalar parameters for PHP native functions is deprecated.

In the tests, a number of errors could be seen along the lines of:
```
date_create_immutable_from_format(): Passing null to parameter #2 ($datetime) of type string is deprecated
```

In this case, both methods in the `Date_Helper` class expect a `string` as input, but it is not verified that they actually receive a `string` and the received parameter is blindly passed on to PHP native functions.
As running `date_create_immutable_from_format()` is useless when no valid date/timestamp is passed, we can short circuit both functions by doing a type check at the start and bowing out early if the received parameter is not a string, or in the case of `format_timestamp()`, not a string nor an integer.

Refs:
* https://wiki.php.net/rfc/deprecate_null_to_scalar_internal_arg
* https://www.php.net/manual/en/datetimeimmutable.createfromformat.php
* https://www.php.net/manual/en/datetime.createfromformat.php