Meta_Surface::for_url(): add extra test for date archive page and improve tests for home page
As the `Meta_Surface::for_url()` method contains a diverging path specifically for date archives, a test needs to be added covering that path.

As the expectation setting in the test function needs changing to accommodate for data archives, we may as well also adjust the "Home page" test cases to be more realistic, i.e. set these to use a `$url` without path component.

This way, the changes being made to accommodate the date archive test, will not hide a PHP 8.1 deprecation notice which is triggered on a `$url` without path component being passed to the `Meta_Surface::is_date_archive_url()` method.

Previously, all test cases would hit the deprecation notice as the third call to `wp_parse_url()` (the call made from within `Meta_Surface::is_date_archive_url()` did not have an expectation nor mocked return, which resulted in that call always returning `null`.

With the changed test logic, the deprecation notice will now only be thrown for the "Home Page" test cases, but that does mean there are tests which cover that particular situation.