WPSEO_Sitemap_Provider_Overlap_Test: stabilize the tests
The `unregister_taxonomy()` function call may not be executed when the tests fails before reaching that point. In this case, that's unlikely as the tests don't contain assertions, but only expectations.
Still, as a matter of best practice and example, I'm moving the `register_taxonomy()` and `unregister_taxonomy()` function calls to the `set_up()` and `tear_down()` fixture methods.