Improve toggle stability in WincherKeyphraseTable
We now show a spinner when a row is loading, rather than faking it being
completed immediately. We also ensure that there's only one active
request to getKeyphrases at the same time. In addition, make sure
requests are fired in the correct order when trackAll is true
(on postPublish).

Also refactor to use a functional component with hooks rather than a
class based one.