Indexable::save(): use `mb_strlen()`
The `Indexable::save()` method uses the `mb_substr()` method to create a substring for the primary keyword, but determines whether or not that substring should be created using the non-binary safe `strlen()` function.

In practice, this means that for primary keywords set in a multi-byte character script, the `mb_substr()` may be called when it is unnecessary as the _byte_ length is more than 191, but the _character_ length is less than 191.

As WP Core polyfills the PHP native `mb_strlen()` function anyway (same as the `mb_substr()` function), it seems more prudent to make sure that these function calls are aligned with each other.

Note: it would be prudent to add an additional test for this, along the same lines as the existing `Indexable_Test::test_save_primary_focus_keyword_truncated()`, but with a text string in a multi-byte character set.
Would love to do so, but I would need some sample text for that. Maybe team Lingo could provide that ?

Refs:
* https://www.php.net/manual/en/function.mb-strlen.php
* https://core.trac.wordpress.org/browser/trunk/src/wp-includes/compat.php#L137