CS/QA: fix inconsistent/incorrect return statement
This method is supposed to return a string (based on the docblock) and the return value is passed on to `WPDB::query()`, which expects a string as well, so probably better to return an empty string than to return boolean `true`.

Alternatively, if there is a good reason to return `true` here, `bool` should be added to the return type and the reason for returning `true` should be documented inline.