Abstract_Aioseo_Settings_Importing_Action_Test: remove redundant test double
The `get_source_option_name()` method in the `Abstract_Aioseo_Settings_Importing_Action_Double` test double class fall through directly to the parent class.
And this same method in the parent class `Abstract_Aioseo_Settings_Importing_Action` is already `public`, so there is no need for a test double.

As for the `transform_robot_setting()` method: that method doesn't even exist in the parent class, so calling the `parent::transform_robot_setting()` method is... well... useless ?