PHP 8.2 | Elementor_Exclude_Post_Types: fix for partially supported callback
`parent::method_name`-like callbacks are only partially supported by PHP. With this mind, they will be deprecated in PHP 8.2, with support being completely removed in PHP 9.0.

Fixed by changing it to a fully supported syntax.

Note: while this issue is not covered by tests, as this is a deprecated class, I'm not going to bother with adding tests.

Refs:
* https://wiki.php.net/rfc/deprecate_partially_supported_callables