Add person social profile save on submit
On submit, two submits are done in parallel.
Whenever one fails, the notification is thrown.
Even though some things might be saved, we treat it as if none has.
This make it simpler. Plus, the settings route already has this behavior in the backend.
When we add frontend validation for the person social profiles, there is no reason left for these errors to occur besides networking.