Use a simple container instead of creating/mocking surfaces with dynamic properties
When mocking away our surfaces_helper, we often combine it with setting dynamic properties.
This is deprecated since PHP 8.2 and will fail tests on PHP >= 8.2.
Also, when we mock away our internals, we hide potential issues in our own code. So,
with these helpers in the testCase, we can remove the need for mocks and actually test our system.