Robots_Txt_Integration: allow for different line endings
Along the same lines as the changes to `Robots_Txt_Helper`, it should be recognized that the `robots.txt` _input_ could be:
* Auto-generated by WP (non-file based).
* File based with unknown line endings.
* Pre-filtered by other plugins which may or may not affect the line endings.

So, this changes the code in the `Robots_Txt_Integration::remove_default_robots()` method to allow for all potential line endings.

Includes a new set of dedicated tests, specifically for the `Robots_Txt_Integration::remove_default_robots()` method, to verify and safeguard the change.