Composer: remove explicit dependencies on the Symfony polyfill libs
The Symfony polyfill libs are not *our* dependency, but a dependency of Guzzle via the Oauth2 library.

The only reason these were in the `require-dev` section, was to prevent these sub-dependencies from being updated to a version incompatible with PHP 5.6, which due to the DI container code, couldn't previously be safeguarded via a `config - platform` setting.

As 1) compatibility with PHP 5.6 is no longer an issue and 2) in the new reality, there will be an appropriate `config - platform` setting, the explicit requirements for these polyfill libs can now be removed and these dependencies (including their sub-dependencies) can now be updated, or in some cases, removed.